+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Fixed
+* fix(perft): enhanced perft testing [[#47](https://github.com/nchandur/ranga/pull/47)]
+* fix(pgn): remove PGN parser [[#46](https://github.com/nchandur/ranga/pull/46)]
+
 ## [v1.12] - 2026-09-19
 
 ### Added